move `getLabelTypeFromFrame` to `languageUtils`
